bugfixing and renaming
